requirements.txt
fastapi
uvicorn
streamlit
scikit-learn
pandas
numpy
sentence-transformers
chromadb
google-genai
python-dotenv
requests
joblib
matplotlib
seaborn
wordcloud
